requirements.txt
google-cloud-storage
python-dotenv
twitter-scraper